app/requirements.txt
Flask==1.1.1
requests==2.21.0
pycryptodome==3.9.4
